move the uva files into cuda directory.
